channel service refactor
