+++ src/mock/v1.2/channel.ts
@@ -1,20 +1,19 @@
-import { Fdc3CommandExecutor1_2 } from './channelService-1_2'
-import { onFdc3Ready } from './mock-functions'
-import { DesktopAgent } from 'fdc3_1_2';
-import { ChannelsAppContext } from '../../test/common/channel-control';
+import { onFdc3Ready } from "./mock-functions";
+import { DesktopAgent } from "fdc3_1_2";
+import { Fdc3CommandExecutor } from "../channel-command";
+import { ChannelService1_2 } from "./support/channel-support-1.2";
+import { ChannelsAppContext } from "../../test/common/channel-control";
 
-declare let fdc3 : DesktopAgent
+declare let fdc3: DesktopAgent;
 
 onFdc3Ready().then(() => {
-    let firedOnce = false;
-    //await commands from App A, then execute commands
-    fdc3.addContextListener(
-        "channelsAppContext",
-        (context: ChannelsAppContext) => {
-            if (firedOnce === false) {
-                firedOnce = true;
-                const commandExecutor = new Fdc3CommandExecutor1_2();
-                commandExecutor.executeCommands(context.commands, context.config);
-            }
-        });
+  let firedOnce = false;
+  //await commands from App A, then execute commands
+  fdc3.addContextListener("channelsAppContext", (context: ChannelsAppContext) => {
+    if (firedOnce === false) {
+      firedOnce = true;
+      const commandExecutor = new Fdc3CommandExecutor();
+      commandExecutor.executeCommands(context.commands, context.config, new ChannelService1_2());
+    }
+  });
 });
